+++ src/docomestria/__init__.py
@@ -18,4 +18,4 @@ __all__ = [
     "pair_labels_to_values",
 ]
 
-__version__ = "0.1.0"
+__version__ = "0.2.0"